refactor client and server.

+++ client/include/client.h
@@ -1,7 +1,7 @@
 #ifndef __CLIENT_H__
 #define __CLIENT_H__
 
-#include "error_code.h"
+#include "../../udp_packet/include/udp_packet.h"
 #include "keylogger.h"
 #include "std_include.h"
 
@@ -13,8 +13,8 @@ typedef struct logger_client {
 
 ErrorCode client_init(logger_client_t *client, uint8_t interface_idx, uint8_t *dest_mac);
 ErrorCode client_start(logger_client_t *client, uint8_t *src_mac, uint8_t *dest_mac, uint8_t *src_ip, uint8_t *dest_ip,
-                       size_t src_port,
-                       size_t dest_port);
+                       uint16_t src_port,
+                       uint16_t dest_port);
 ErrorCode client_destroy(logger_client_t *client);
 
 #endif
